zoefit/requirements.txt
Django>=6.0.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
Pillow>=10.0.0
python-dotenv>=1.0.0

# AI/ML Libraries
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
google-generativeai>=0.3.0

# Data Processing
requests>=2.31.0

